refactor: move sidebar panels into features
This removes 2 levels of nested folders for 5 features.
Will simplify fs navigating a lot.
