+++ apps/builder/app/builder/features/pages/image-info.tsx
