+++ apps/builder/app/builder/features/pages/page-utils.test.ts
@@ -14,7 +14,6 @@ import {
   registerFolderChildMutable,
   reparentOrphansMutable,
   filterSelfAndChildren,
-  getExistingRoutePaths,
   $pageRootScope,
   isPathAvailable,
 } from "./page-utils";
@@ -379,35 +378,6 @@ describe("filterSelfAndChildren", () => {
   test("filter self and child folders", () => {
     const result = filterSelfAndChildren("1", folders);
     expect(result).toEqual([folders[2]]);
-  });
-});
-
-describe("getExistingRoutePaths", () => {
-  const { pages } = createPages();
-
-  test("gets all the route paths that exists in the project", () => {
-    pages.pages.push({
-      id: "pageId",
-      meta: {},
-      name: "Page",
-      path: "/page",
-      rootInstanceId: "rootInstanceId",
-      systemDataSourceId: "systemDataSourceId",
-      title: `"Page"`,
-    });
-
-    pages.pages.push({
-      id: "blogId",
-      meta: {},
-      name: "Blog",
-      path: "/blog/:id",
-      rootInstanceId: "rootInstanceId",
-      systemDataSourceId: "systemDataSourceId",
-      title: `"Blog"`,
-    });
-
-    const result = getExistingRoutePaths(pages);
-    expect(Array.from(result)).toEqual(["/page", "/blog/:id"]);
   });
 });
