+++ apps/builder/app/builder/shared/extended-sidebar-panel.tsx
